requirements.txt
gradio
python-dotenv
pandas
loguru
psutil
httpx       
pybreaker     


# Document parsing dependencies
textract
pytesseract
tenacity
python-docx
pypdf
pymupdf    # import fitz
PyPDF2
pillow
ebooklib
beautifulsoup4
mimetype
# PyPDF2==3.0.1

# AI dependencies
openai
tiktoken
anthropic
google-cloud-aiplatform
google-auth

# RENDER.io dependencies
fastapi
uvicorn
celery
gevent==23.9.1
redis 
websockets 

# AWS dependencies
boto3
botocore

# Supabase & PG dependencies
supabase 
postgrest-py
asyncpg    

# RAG dependencies
langchain 
langchain-community
langchain-openai
promptic